delete/move some manual types
