+++ frontend/src/employee-frontend/types/absence.ts
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2017-2022 City of Espoo
+// SPDX-FileCopyrightText: 2017-2024 City of Espoo
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
@@ -6,20 +6,12 @@ import {
   AbsenceCategory,
   AbsenceType
 } from 'lib-common/generated/api-types/absence'
-import LocalDate from 'lib-common/local-date'
-import { UUID } from 'lib-common/types'
 
-export const defaultAbsenceType = 'SICKLEAVE'
-export const defaultAbsenceCategories: AbsenceCategory[] = []
-
-export const absenceCategories: AbsenceCategory[] = ['NONBILLABLE', 'BILLABLE']
-
-type AbsenceTypeWithBackupCare = AbsenceType | 'TEMPORARY_RELOCATION'
-
-export interface CellPart {
-  childId: UUID
-  date: LocalDate
-  absenceType: AbsenceTypeWithBackupCare | undefined
-  category: AbsenceCategory
-  position: 'left' | 'right'
-}
+export type AbsenceUpdate =
+  | {
+      type: 'absence'
+      absenceType: AbsenceType
+      absenceCategories: AbsenceCategory[]
+    }
+  | { type: 'noAbsence'; absenceCategories: AbsenceCategory[] }
+  | { type: 'missingHolidayReservation' }
